+++ api/src/main/java/io/strimzi/api/kafka/model/KafkaExporterSpec.java
@@ -29,7 +29,7 @@ import java.util.Map;
     "image", "groupRegex", "topicRegex", 
     "groupExcludeRegex", "topicExcludeRegex",
     "resources", "logging",
-    "enableSaramaLogging", "template"})
+    "enableSaramaLogging", "showAllOffsets", "template"})
 @EqualsAndHashCode
 public class KafkaExporterSpec implements HasLivenessProbe, HasReadinessProbe, UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
@@ -45,6 +45,7 @@ public class KafkaExporterSpec implements HasLivenessProbe, HasReadinessProbe, U
     private Probe readinessProbe;
     private String logging = "info";
     private boolean enableSaramaLogging;
+    private boolean showAllOffsets = true;
     private KafkaExporterTemplate template;
     private Map<String, Object> additionalProperties = new HashMap<>(0);
 
@@ -110,6 +111,16 @@ public class KafkaExporterSpec implements HasLivenessProbe, HasReadinessProbe, U
 
     public void setEnableSaramaLogging(boolean enableSaramaLogging) {
         this.enableSaramaLogging = enableSaramaLogging;
+    }
+
+    @Description("Whether show the offset/lag for all consumer group, otherwise, only show connected consumer groups.")
+    @JsonInclude(JsonInclude.Include.NON_DEFAULT)
+    public boolean getShowAllOffsets() {
+        return showAllOffsets;
+    }
+
+    public void setShowAllOffsets(boolean showAllOffsets) {
+        this.showAllOffsets = showAllOffsets;
     }
 
     @Description("Only log messages with the given severity or above. " +